refactor chat-math
