--- src/app/features/chat-math-solve/application/services/chat.service.ts
@@ -1,12 +1,0 @@
-import { Injectable, InjectionToken } from '@angular/core';
-import { SendMessageUseCase } from '@features/chat-math-solve/application/uses-cases/send-message.use-case';
-import { Message } from '@features/chat-math-solve/domain/entities/message.entity';
-
-@Injectable({ providedIn: 'root' })
-export class ChatService {
-    constructor(private readonly _sendMessageUseCase: SendMessageUseCase) {}
-
-    sendMessage(message: Message) {
-        return this._sendMessageUseCase.execute(message);
-    }
-}
