--- src/app/features/chat-math-solve/presentation/components/chat-item/chat-item.component.scss
@@ -1,71 +1,0 @@
-@use "mixins";
-
-.chat {
-    &__item {
-        color: var(--color-text);
-        --background: transparent;
-        --padding-start: 0;
-        --padding-end: 0;
-        --inner-padding-end: 0; // Remove default ion-item padding
-        --min-height: auto;
-        margin-bottom: var(--spacing-md); // Increased spacing between messages
-        display: flex; // Enables flex alignment for bubble
-        width: 100%;
-
-        &--user {
-            justify-content: end;
-            .chat__bubble {
-                margin-left: auto;
-            }
-        }
-
-        &--ai {
-            justify-content: start;
-        }
-    }
-
-    &__bubble {
-        margin: 0;
-        font-size: var(--normal-font-size);
-        line-height: 1.5;
-        white-space: pre-wrap; // Preserve line breaks in messages
-        width: 80%; // Allow slightly wider bubbles
-        word-wrap: break-word; // Ensure long words break
-        border-radius: var(--border-radius-md);
-        padding: var(--spacing-sm) var(--size-sm); // Adjusted padding
-
-        &-text {
-            font-size: var(--normal-font-size);
-            line-height: 1.5;
-            white-space: pre-wrap; // Preserve line breaks in messages
-        }
-
-        &--user {
-            @include mixins.fadeUpChat();
-            background-color: var(--color-primary);
-            border: 1px solid rgba(255, 255, 255, 0.1);
-            color: var(--color-white);
-        }
-
-        &--ai {
-            @include mixins.fadeInChat();
-            background: var(--gradient-bg-ai);
-            border: 1px solid rgba(255, 255, 255, 0.1);
-            color: var(--color-text);
-        }
-    }
-
-    &__images {
-        display: flex;
-        gap: var(--spacing-sm);
-        margin-bottom: var(--spacing-xs);
-    }
-
-    &__image {
-        width: 4rem;
-        height: 4rem;
-        object-fit: cover;
-        overflow: hidden;
-        border-radius: var(--border-radius-sm);
-    }
-}
