--- src/app/features/chat-math-solve/presentation/components/chat-item/chat-item.component.ts
@@ -1,33 +1,0 @@
-import { Component, CUSTOM_ELEMENTS_SCHEMA, input, ElementRef, ViewChild } from '@angular/core';
-import { IonItem, IonImg } from '@ionic/angular/standalone';
-import { CommonModule } from '@angular/common';
-import { MessageState } from '@features/chat-math-solve/application/states/interfaces/chat-math.state.interface';
-import { ChatActionsComponent } from '../chat-actions/chat-actions.component';
-
-@Component({
-    selector: 'app-chat-item',
-    template: `
-        <ion-item class="chat__item" [ngClass]="'chat__item--' + message().role">
-            <div class="chat__bubble" [ngClass]="'chat__bubble--' + message().role">
-                @if (message().images?.length) {
-                <div class="chat__images">
-                    @for (image of message().images; track $index) {
-                    <ion-img [src]="image.webPath" class="chat__image"></ion-img>
-                    }
-                </div>
-                }
-                <p class="chat__bubble-text" #messageContent>{{ message().content }}</p>
-                @if (message().role === 'ai') {
-                <app-chat-actions [content]="message().content"></app-chat-actions>
-                }
-            </div>
-        </ion-item>
-    `,
-    styleUrls: ['./chat-item.component.scss'],
-    imports: [IonItem, IonImg, CommonModule, ChatActionsComponent],
-    schemas: [CUSTOM_ELEMENTS_SCHEMA],
-})
-export class ChatItemComponent {
-    @ViewChild('messageContent') messageContent!: ElementRef<HTMLParagraphElement>;
-    message = input.required<MessageState>();
-}
