--- src/app/features/chat-math-solve/presentation/components/chat-list/chat-list.component.scss
@@ -1,4 +1,0 @@
-.chat__list {
-    background-color: transparent;
-    padding: 0;
-}
