--- src/app/features/chat-math-solve/presentation/components/chat-list/chat-list.component.ts
@@ -1,31 +1,0 @@
-import { Component, Inject, OnInit } from '@angular/core';
-import { IonList } from '@ionic/angular/standalone';
-import { MESSAGES_STATE } from '@features/chat-math-solve/application/states/chat-math.state';
-import { StateStorage } from '@shared/storage/interfaces/state-storage.interface';
-import { MessageState } from '@features/chat-math-solve/application/states/interfaces/chat-math.state.interface';
-import { ChatItemComponent } from '../chat-item/chat-item.component';
-import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { ChatWelcomeComponent } from '../chat-welcome/chat-welcome.component';
-
-@Component({
-    selector: 'app-chat-list',
-    template: `
-        <app-chat-welcome></app-chat-welcome>
-        <ion-list class="chat__list" lines="none">
-            @for (message of $messages(); track $index) {
-            <app-chat-item [message]="message"></app-chat-item>
-            }
-        </ion-list>
-    `,
-    styleUrls: ['./chat-list.component.scss'],
-    imports: [IonList, ChatItemComponent, CommonModule, ChatWelcomeComponent],
-    schemas: [CUSTOM_ELEMENTS_SCHEMA],
-})
-export class ChatListComponent implements OnInit {
-    $messages = this._messagesState.$state;
-
-    constructor(@Inject(MESSAGES_STATE) private _messagesState: StateStorage<Array<MessageState>>) {}
-
-    ngOnInit() {}
-}
