--- src/app/features/chat-math-solve/presentation/components/chat-welcome/chat-welcome.component.scss
@@ -1,3 +1,0 @@
-.chat__welcome {
-    --background: transparent;
-}
