--- src/app/features/chat-math-solve/presentation/pages/chat-math-solve/chat-resolve-math.page.html
@@ -1,10 +1,0 @@
-<app-header title="Ai Assistant" [showBackButton]="true"></app-header>
-
-<ion-content class="chat-resolve-math__content" [fullscreen]="true">
-    <app-chat-list></app-chat-list>
-    @if (isLoading()) {
-    <app-typing-loading></app-typing-loading>
-    }
-</ion-content>
-
-<app-footer (onSendMessage)="sendMessage($event)"></app-footer>
