--- src/app/features/chat-math-solve/presentation/pages/chat-math-solve/chat-resolve-math.page.scss
@@ -1,6 +1,0 @@
-ion-content {
-    --padding-start: var(--spacing-md);
-    --padding-end: var(--spacing-md);
-    --padding-top: var(--spacing-md);
-    --padding-bottom: var(--spacing-md);
-}
