+++ src/app/features/chat-math/domain/entities/image.entity.ts
