+++ src/app/features/chat-math/domain/entities/message.entity.ts
