+++ src/app/features/chat-math/domain/factories/message.factory.ts
