+++ src/app/features/chat-math/presentation/components/chat-list/chat-list.component.scss
@@ -1,0 +1,3 @@
+.chat__list {
+    background-color: transparent;
+}
