+++ src/app/features/chat-math/presentation/components/footer-preview/footer-preview.component.scss
