+++ src/app/features/chat-math/presentation/components/footer/footer.component.scss
