+++ src/app/features/chat-math/presentation/components/modal-image-preview/modal-image-preview.component.html
