+++ src/app/shared/utils/constants/app.constants.ts
@@ -1,0 +1,1 @@
+export const DURATION_MS_TOAST = 2000;
