+++ src/app/shared/utils/constants/chat-math.constants.ts
@@ -1,0 +1,4 @@
+export const TEXT_WEB_SHARE_API_NOT_AVAILABLE = 'Web Share API no disponible';
+export const TEXT_COPY_TO_CLIPBOARD = 'Texto copiado al portapapeles';
+export const TEXT_DOWNLOAD_AS_PDF = 'Descargar como PDF';
+export const TEXT_CHAT_WELCOME = '¡Hola! Soy tu asistente EduAiTutor. Envíame tu ejercicio y te ayudo paso a paso. 😊';
