--- src/app/shared/utils/constants/contant.ts
@@ -1,5 +1,0 @@
-export const DURATION_TOAST = 2000;
-export const TEXT_WEB_SHARE_API_NOT_AVAILABLE = 'Web Share API no disponible';
-export const TEXT_COPY_TO_CLIPBOARD = 'Texto copiado al portapapeles';
-export const TEXT_DOWNLOAD_AS_PDF = 'Descargar como PDF';
-export const TEXT_CHAT_WELCOME = '¡Hola! Soy tu asistente EduAiTutor. Envíame tu ejercicio y te ayudo paso a paso. 😊';
