+++ src/app/shared/utils/constants/index.ts
@@ -1,0 +1,2 @@
+export * from './chat-math.constants';
+export * from './app.constants';
