+++ src/main.ts
@@ -9,6 +9,7 @@ import { firebaseProviders } from './app/shared/config/firebase.config';
 import { enableProdMode, importProvidersFrom } from '@angular/core';
 import { environment } from '@envs/environment';
 import { appInitializerProviders } from './app/shared/config/app-initializer.config';
+import { CHAT_PROVIDER } from '@features/chat-math/infrastructure/providers/provider';
 
 if (environment.production) {
     enableProdMode();
@@ -24,5 +25,6 @@ bootstrapApplication(AppComponent, {
         //TODO: Firebase config start
         ...firebaseProviders,
         importProvidersFrom(),
+        ...CHAT_PROVIDER,
     ],
 });
